Introduced sorted dictionary to have scripts executed in an ordered fashion.

+++ utils/create_version_table.sql
@@ -1,4 +1,4 @@
 CREATE TABLE databaseversion(
-    Revision int NOT NULL,
-    Applied TIMESTAMP WITH TIME ZONE NOT NULL,
+    Revision BIGINT NOT NULL,
+    Applied TIMESTAMP WITH TIME ZONE NOT NULL
     )
